requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
chromadb==0.4.18
pypdf==3.17.4
python-dotenv==1.0.0
typing-extensions==4.8.0